api/requirements.txt
transformers
pillow
torch
numpy 